requirements.txt
azure==1.0.3
azure-common==1.1.4
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==1.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==1.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
futures==3.0.5
python-dateutil==2.5.3
python-ntlm3==1.0.2
pywinrm==0.2.1
requests==2.11.1
requests-ntlm==0.3.0
six==1.10.0
xmltodict==0.10.2
